+++ common/src/main/java/club/iananderson/seasonhud/client/gui/screens/MainConfigScreen.java
@@ -107,12 +107,18 @@ public class MainConfigScreen extends SeasonHudScreen {
                 Component.translatable("menu.seasonhud.main.enableMod.button"), (b, val) -> enableMod = val);
 
     int row = 0;
-    seasonButton = MenuButton.builder(MenuButtons.SEASON, b -> SeasonOptionsScreen.getInstance(this).open())
+    seasonButton = MenuButton.builder(MenuButtons.SEASON, b -> {
+          this.saveConfig();
+          SeasonOptionsScreen.getInstance(this).open();
+        })
         .withTooltip(Tooltip.create(Component.translatable("menu.seasonhud.main.season.tooltip")))
         .withPos(leftButtonX, (buttonStartY + (row * yOffset))).withWidth(BUTTON_WIDTH)
         .build();
 
-    colorButton = MenuButton.builder(MenuButtons.COLORS, b -> ColorScreen.getInstance(this).open())
+    colorButton = MenuButton.builder(MenuButtons.COLORS, b -> {
+          this.saveConfig();
+          ColorScreen.getInstance(this).open();
+        })
         .withTooltip(Tooltip.create(Component.translatable("menu.seasonhud.main.color.tooltip")))
         .withPos(rightButtonX, (buttonStartY + (row * yOffset))).withWidth(BUTTON_WIDTH)
         .build();