+++ common/src/main/java/club/iananderson/seasonhud/client/gui/screens/SeasonHudScreen.java
@@ -5,12 +5,14 @@ import club.iananderson.seasonhud.client.gui.components.buttons.MenuButton.MenuB
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Stream;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.ComponentPath;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.AbstractWidget;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
+import net.minecraftforge.common.ForgeConfigSpec.ConfigValue;
 import org.jetbrains.annotations.NotNull;
 
 public class SeasonHudScreen extends Screen {
@@ -28,6 +30,8 @@ public class SeasonHudScreen extends Screen {
   public int row;
   public int buttonStartY = MENU_PADDING;
   public int yOffset = BUTTON_HEIGHT + BUTTON_PADDING;
+  protected boolean hasPendingChanges;
+  protected List<ConfigValue<?>> configOptions = new ArrayList<>();
 
   public SeasonHudScreen(Screen parentScreen, Component title) {
     super(title);
@@ -46,12 +50,29 @@ public class SeasonHudScreen extends Screen {
     return true;
   }
 
-  public void onDone() {
-    Minecraft.getInstance().setScreen(this.parentScreen);
+  @Override
+  public boolean shouldCloseOnEsc() {
+    return !this.hasPendingChanges;
+  }
+
+  private Stream<ConfigValue<?>> getAllSettings() {
+    return this.configOptions.stream();
+  }
+
+  private void undoChanges() {
+    this.getAllSettings().forEach(ConfigValue::clearCache);
+  }
+
+  public void saveConfig() {
+
   }
 
   @Override
   public void onClose() {
+    Minecraft.getInstance().setScreen(this.parentScreen);
+  }
+
+  public void onDone() {
     Minecraft.getInstance().setScreen(this.parentScreen);
   }
 
@@ -63,10 +84,8 @@ public class SeasonHudScreen extends Screen {
 
   }
 
-  protected void rebuildWidgets() {
-    this.clearWidgets();
-    this.clearFocus();
-    this.init();
+  public void rebuildUI() {
+    this.rebuildWidgets();
   }
 
   @Override
@@ -80,20 +99,20 @@ public class SeasonHudScreen extends Screen {
 
   @Override
   public void init() {
-    this.widgets.clear();
     super.init();
+    this.widgets.clear();
     leftButtonX = (this.width / 2) - (BUTTON_WIDTH + BUTTON_PADDING);
     rightButtonX = (this.width / 2) + BUTTON_PADDING;
-
-    doneButton = MenuButton.builder(MenuButtons.DONE, press -> this.onDone())
-        .withPos(rightButtonX, (this.height - MenuButton.DEFAULT_HEIGHT - BUTTON_PADDING))
-        .build();
 
     cancelButton = MenuButton.builder(MenuButtons.CANCEL, press -> this.onClose())
         .withPos((this.width / 2) - (MenuButton.DEFAULT_WIDTH + BUTTON_PADDING),
                  (this.height - MenuButton.DEFAULT_HEIGHT - BUTTON_PADDING))
         .build();
 
-    this.widgets.addAll(Arrays.asList(doneButton, cancelButton));
+    doneButton = MenuButton.builder(MenuButtons.DONE, press -> this.onDone())
+        .withPos(rightButtonX, (this.height - MenuButton.DEFAULT_HEIGHT - BUTTON_PADDING))
+        .build();
+
+    this.widgets.addAll(Arrays.asList(cancelButton, doneButton));
   }
 }