+++ common/src/main/java/club/iananderson/seasonhud/client/overlays/SeasonHUDOverlayCommon.java
@@ -24,34 +24,39 @@ public class SeasonHUDOverlayCommon {
     double scale = Config.getHudScale();
     int DEFAULT_X_OFFSET = Config.DEFAULT_X_OFFSET;
     int DEFAULT_Y_OFFSET = Config.DEFAULT_Y_OFFSET;
-    int stringWidth = mc.font.width(seasonCombined);
-    int stringHeight = mc.font.lineHeight;
+    int stringWidth = (int) (mc.font.width(seasonCombined) * scale);
+    int stringHeight = (int) (mc.font.lineHeight * scale);
 
     if (Common.drawDefaultHud() && Common.vanillaShouldDrawHud() && Calendar.validNeedCalendar()) {
       switch (Config.getHudLocation()) {
         case TOP_LEFT:
-          x = xOffset;
-          y = yOffset;
+          x = DEFAULT_X_OFFSET;
+          y = DEFAULT_Y_OFFSET;
           break;
 
         case TOP_CENTER:
-          x = (screenWidth / 2) - (stringWidth / 2);
+          x = (int) ((((double) screenWidth / 2) - ((double) stringWidth / 2)) / scale);
           y = DEFAULT_Y_OFFSET;
           break;
 
         case TOP_RIGHT:
-          x = screenWidth - stringWidth - DEFAULT_X_OFFSET;
+          x = (int) ((screenWidth - stringWidth - DEFAULT_X_OFFSET) / scale);
           y = DEFAULT_Y_OFFSET;
           break;
 
         case BOTTOM_LEFT:
           x = DEFAULT_X_OFFSET;
-          y = screenHeight - stringHeight - DEFAULT_Y_OFFSET;
+          y = (int) (((screenHeight - stringHeight - DEFAULT_Y_OFFSET)) / scale);
           break;
 
         case BOTTOM_RIGHT:
-          x = screenWidth - stringWidth - DEFAULT_X_OFFSET;
-          y = screenHeight - stringHeight - DEFAULT_Y_OFFSET;
+          x = (int) (((screenWidth - stringWidth - DEFAULT_X_OFFSET)) / scale);
+          y = (int) (((screenHeight - stringHeight - DEFAULT_Y_OFFSET)) / scale);
+          break;
+
+        case CUSTOM:
+          x = (int) xOffset;
+          y = (int) yOffset;
           break;
       }