+++ CHANGELOG.md
@@ -4,18 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
-## [1.13.4] - 2025-05-27
+## [1.13.4] - 2025-06-22
 
 ### Added
 
 - Added a scaling option for the default hud
+- Added a "Custom" HUD location
+    - The custom X and Y offsets will now only take effect if using this.
 
 ## [1.13.3] - 2025-05-06
 
 ### Changed
 
 - Updated to support new XaerosMinimap InfoDisplay changes.
-  - Minimum version is now **25.2.0**
+    - Minimum version is now **25.2.0**
 
 ## [1.13.2] - 2025-04-13